week 2/Basic Prompt/requirements.txt
bcrypt==4.1.2
pyjwt==2.8.0
python-dotenv==1.0.1 